requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
boto3==1.35.99
botocore==1.35.99
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dj-logging==2.0.1
Django==5.1.7
django-celery-beat==2.8.0
django-ckeditor==6.7.2
django-filter==25.1
django-js-asset==3.1.2
django-phonenumber-field==8.1.0
django-redis==5.4.0
django-restframework==0.0.1
django-storages==1.14.5
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
gunicorn==23.0.0
hiredis==3.1.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kavenegar==1.1.2
kombu==5.5.3
packaging==25.0
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
